requirements.txt
psycopg2==2.6.1
Markdown==2.6.2
Django==1.8.5
django-recaptcha==1.0.4
